Places scatter subplot in its own module
